Moves 'proxy_request' function to own module (proxy.py)
